Security: if credential is stored in php, ignore posted
Bugfix: app sends a test_connection, if hardcoded_credentials is returned, it uses the returned data in password, as password,
and the user can't change it in the app, with previus security update, the field passwords now only says yes or no,
so the app sends the word yes as password back to the server on the 2nd request.
by ignoring the credential in the posted data, thats no longer a problem.